Remove tracked binaries, fix all clippy warnings
Remove libfroztlib.dylib and frozt-party from git tracking and add
gitignore rules for build artifacts. Delete stale pre-rename
libfrostlib.dylib. Fix 12 clippy warnings: assign ops, unnecessary
clone on Copy type, redundant cast, and const thread_local init.

+++ frozt-wasm/src/codec.rs
@@ -70,7 +70,7 @@ pub fn wasm_encode_map(entries: JsValue) -> Result<Vec<u8>, JsError> {
     let arr = js_sys::Array::from(&entries);
     let len = arr.length();
     let mut buf = Vec::new();
-    buf.extend_from_slice(&(len as u32).to_le_bytes());
+    buf.extend_from_slice(&len.to_le_bytes());
     for i in 0..len {
         let entry = arr.get(i);
         let id_val = js_sys::Reflect::get(&entry, &"id".into())
